refactor: extract login form and remember username hook
